fix: address PR review feedback
- Remove dead validator functions (isValidDamageMessage, isValidProjectileMessage)
- Extract shared applyDamageToTarget to combatUtils.ts
- Move DEFAULT_PROJECTILE_RADIUS to shared/constants.ts
- Replace onStateChange with per-projectile listeners in OnlineGameMode
- Change lastProcessedSeq from int16 to uint32 for seq overflow safety
- Optimize InputBuffer.add() spread to push
- Add design comments for one-input-per-tick and attackTargetId convention
